requirements.txt
certifi==2019.3.9
chardet==3.0.4
Click==7.0
Flask==1.0.2
gcloud==0.18.3
googleapis-common-protos==1.5.9
gunicorn==19.9.0
httplib2==0.12.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
jws==0.1.3
MarkupSafe==1.1.1
oauth2client==4.1.2
protobuf==3.7.1
pyasn1==0.4.5
pyasn1-modules==0.2.4
pycryptodome==3.8.1
Pyrebase4==4.3.0
python-jwt==2.0.1
requests==2.21.0
requests-toolbelt==0.9.1
rsa==4.0
six==1.12.0
urllib3==1.24.2
Werkzeug==0.15.2
